pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sono-eval"
version = "0.1.1"
description = "Explainable multi-path developer assessment system with semantic tagging and analytics"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Sono-Eval Team"}
]
keywords = ["assessment", "evaluation", "developer", "analytics", "ml", "tagging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "peft>=0.7.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.25.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
    "celery>=5.3.0",
    "kombu>=5.3.0",
    "council-ai>=2.0.0",
    "textstat>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]
memu = [
    "memu-py>=1.0.0",
]
all = [
    "memu-py>=1.0.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.scripts]
sono-eval = "sono_eval.cli.main:cli"

[project.urls]
Homepage = "https://github.com/doronpers/sono-eval"
Repository = "https://github.com/doronpers/sono-eval"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']
extend-exclude = '''
/(
  ^venv.*
  | ^\.venv
)/
'''

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/sono_eval --cov-report=html --cov-report=term"

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["venv*", "env*", ".venv*"]

[tool.bandit]
exclude_dirs = ["tests", "venv", "venv-test", ".venv"]

[tool.pyrefly]
project-excludes = [
    "build",
    "dist",
    "models",
    "data",
    "venv",
    ".venv",
    "**/__pycache__"
]
